# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SMS ANGOLA",
        "version": "1.0"
    },
    "paths": {
...
        "/api/senders": {
            "get": {
                "tags": [
                    "Senders"
                ],
                "description": "Liste des expéditeurs.",
                "operationId": "listSender",
                "responses": {
                    "200": {
                        "description": "Liste des expéditeurs."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Senders"
                ],
                "description": "Enregistrement d'un expéditeur.",
                "operationId": "storeSender",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label",
                                    "bydefault"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string"
                                    },
                                    "bydefault": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expéditeur enregisté avec succès."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/senders/{uid}": {
            "get": {
                "tags": [
                    "Senders"
                ],
                "description": "Détail d'un expéditeur.",
                "operationId": "showSender",
                "responses": {
                    "200": {
                        "description": "Détail d'un expéditeur."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Senders"
                ],
                "description": "Modification d'un expéditeur.",
                "operationId": "editSender",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label",
                                    "bydefault"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string"
                                    },
                                    "bydefault": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expéditeur modifié avec succès."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Senders"
                ],
                "description": "Suppression d'un expéditeur.",
                "operationId": "deleteSender",
                "responses": {
                    "201": {
                        "description": "Expéditeur supprimé avec succès."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/senders/validated": {
            "post": {
                "tags": [
                    "Senders"
                ],
                "description": "Validation d'un expéditeur.",
                "operationId": "validSender",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "uid"
                                ],
                                "properties": {
                                    "uid": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expéditeur validé avec succès."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
}